requirements.txt
flask==3.1.1
requests==2.32.3
python-dotenv==1.1.0
gunicorn==23.0.0
stripe==11.4.1
